# Support CONCEPTQL_AVOID_CTES=true to avoid creating CTEs where possible