--- a/lib/conceptql.rb
+++ b/lib/conceptql.rb
@@ -16,6 +16,10 @@ Dir.glob(File.dirname(__FILE__) + "/../lib/conceptql/query_modifiers/**/*.rb").e
 end
 
 module ConceptQL
+  def self.avoid_ctes?
+    ENV['CONCEPTQL_AVOID_CTES'] == 'true'
+  end
+
   def self.metadata(opts = {})
     {
       categories: categories,