CTEs are used in 4 places on PostgreSQL (1 additional place on Impala,
which I didn't modify):

* co_reported operator
* occurance operator (2 places)
* scope (for implementing recall operator)

Unfortunately, this approach does not help performance in the tests
when enabled:

co_reported: 15.2 seconds -> 16.6 seconds
occurrence: 50 seconds -> 152 seconds
recall: 1.4 seconds -> 2.5 seconds